fixed mongo queries + restructured
