+++ server/mongodb/actions/Chapter.ts
@@ -17,9 +17,3 @@ export async function chapterUpdateChapter(
 
   return updatedChapter;
 }
-
-export async function adminGetChapters() {
-  await dbConnect();
-
-  return ChapterModel.find();
-}
