+++ server/mongodb/actions/Nonprofit.ts
@@ -30,9 +30,3 @@ export async function nonprofitUpdateNonprofit(
 
   return nonprofit;
 }
-
-export async function adminGetNonprofits() {
-  await dbConnect();
-
-  return NonprofitModel.find();
-}
