+++ server/mongodb/actions/User.ts
@@ -74,9 +74,3 @@ export async function nonprofitUpdateUser(
 
   return user;
 }
-
-export async function adminGetChapterUsers(chapterId: Types.ObjectId) {
-  await dbConnect();
-
-  return UserModel.find({ chapter: chapterId });
-}
